Add lib to preflight
