+++ cmd/check_operator_test.go
@@ -1,13 +1,7 @@
 package cmd
 
 import (
-	"context"
 	"os"
-
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/engine"
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/formatters"
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/policy"
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/runtime"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -113,64 +107,6 @@ var _ = Describe("Check Operator", func() {
 				err := ensureIndexImageConfigIsSet()
 				Expect(err).To(HaveOccurred())
 			})
-		})
-	})
-
-	Context("When instantiating a checkOperatorRunner", func() {
-		var cfg *runtime.Config
-		BeforeEach(func() {
-			cfg = &runtime.Config{
-				Image:          "quay.io/example/foo:latest",
-				ResponseFormat: formatters.DefaultFormat,
-			}
-		})
-
-		Context("with a valid policy formatter", func() {
-			It("should return with no error, and the appropriate formatter", func() {
-				cfg.ResponseFormat = "xml"
-				runner, err := newCheckOperatorRunner(context.TODO(), cfg)
-				Expect(err).ToNot(HaveOccurred())
-				expectedFormatter, err := formatters.NewByName(cfg.ResponseFormat)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(runner.formatter.PrettyName()).To(Equal(expectedFormatter.PrettyName()))
-			})
-		})
-
-		Context("with an invalid policy formatter", func() {
-			It("should return an error", func() {
-				cfg.ResponseFormat = "foo"
-				_, err := newCheckOperatorRunner(context.TODO(), cfg)
-				Expect(err).To(HaveOccurred())
-			})
-		})
-
-		Context("with an invalid policy definition", func() {
-			It("should return the container policy engine anyway", func() {
-				cfg.Policy = "badpolicy"
-				beforeCfg := *cfg
-				runner, err := newCheckOperatorRunner(context.TODO(), cfg)
-				Expect(err).ToNot(HaveOccurred())
-
-				_, err = engine.NewForConfig(context.TODO(), cfg.ReadOnly())
-				Expect(runner.cfg.Policy).ToNot(Equal(beforeCfg.Policy))
-				Expect(runner.cfg.Policy).To(Equal(policy.PolicyOperator))
-				Expect(err).ToNot(HaveOccurred())
-			})
-		})
-
-		Context("with an invalid formatter definition", func() {
-			It("should return an error", func() {
-				cfg.ResponseFormat = "foo"
-				_, err := newCheckOperatorRunner(context.TODO(), cfg)
-				Expect(err).To(HaveOccurred())
-			})
-		})
-
-		It("should contain a ResultWriterFile resultWriter", func() {
-			runner, err := newCheckOperatorRunner(context.TODO(), cfg)
-			Expect(err).ToNot(HaveOccurred())
-			_, rwIsExpectedType := runner.rw.(*runtime.ResultWriterFile)
-			Expect(rwIsExpectedType).To(BeTrue())
 		})
 	})
