+++ cmd/check_test.go
@@ -1,12 +1,10 @@
 package cmd
 
 import (
-	"context"
 	"os"
-	"path/filepath"
 
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/artifacts"
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/runtime"
+	"github.com/redhat-openshift-ecosystem/openshift-preflight/internal/lib"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -46,7 +44,7 @@ var _ = Describe("cmd package check command", func() {
 		Context("Regular Connect URL", func() {
 			It("should return a URL with just a project ID", func() {
 				expected := "https://connect.redhat.com/projects/this-is-my-project-id"
-				actual := buildConnectURL(projectID)
+				actual := lib.BuildConnectURL(projectID)
 				Expect(expected).To(Equal(actual))
 			})
 		})
@@ -56,7 +54,7 @@ var _ = Describe("cmd package check command", func() {
 			})
 			It("should return a URL for QA", func() {
 				expected := "https://connect.qa.redhat.com/projects/this-is-my-project-id"
-				actual := buildConnectURL(projectID)
+				actual := lib.BuildConnectURL(projectID)
 				Expect(expected).To(Equal(actual))
 			})
 		})
@@ -66,7 +64,7 @@ var _ = Describe("cmd package check command", func() {
 			})
 			It("should return a URL for UAT", func() {
 				expected := "https://connect.uat.redhat.com/projects/this-is-my-project-id/images/my-image-id/scan-results"
-				actual := buildScanResultsURL(projectID, imageID)
+				actual := lib.BuildScanResultsURL(projectID, imageID)
 				Expect(expected).To(Equal(actual))
 			})
 		})
@@ -76,7 +74,7 @@ var _ = Describe("cmd package check command", func() {
 			})
 			It("should return a URL for QA", func() {
 				expected := "https://connect.qa.redhat.com/projects/this-is-my-project-id/overview"
-				actual := buildOverviewURL(projectID)
+				actual := lib.BuildOverviewURL(projectID)
 				Expect(expected).To(Equal(actual))
 			})
 		})
@@ -86,39 +84,13 @@ var _ = Describe("cmd package check command", func() {
 			})
 			It("should return a Prod overview URL", func() {
 				expected := "https://connect.redhat.com/projects/this-is-my-project-id/overview"
-				actual := buildOverviewURL(projectID)
+				actual := lib.BuildOverviewURL(projectID)
 				Expect(expected).To(Equal(actual))
 			})
 			It("should return a Prod scan URL", func() {
 				expected := "https://connect.redhat.com/projects/this-is-my-project-id/images/my-image-id/scan-results"
-				actual := buildScanResultsURL(projectID, imageID)
+				actual := lib.BuildScanResultsURL(projectID, imageID)
 				Expect(expected).To(Equal(actual))
-			})
-		})
-	})
-
-	Describe("JUnit", func() {
-		var results *runtime.Results
-		var junitfile string
-
-		BeforeEach(func() {
-			results = &runtime.Results{
-				TestedImage:       "registry.example.com/example/image:0.0.1",
-				PassedOverall:     true,
-				TestedOn:          runtime.UnknownOpenshiftClusterVersion(),
-				CertificationHash: "sha256:deadb33f",
-				Passed:            []runtime.Result{},
-				Failed:            []runtime.Result{},
-				Errors:            []runtime.Result{},
-			}
-			junitfile = filepath.Join(artifacts.Path(), "results-junit.xml")
-		})
-
-		When("The additional JUnitXML results file is requested", func() {
-			It("should be written to the artifacts directory without error", func() {
-				Expect(writeJUnit(context.TODO(), *results)).To(Succeed())
-				_, err := os.Stat(junitfile)
-				Expect(err).ToNot(HaveOccurred())
 			})
 		})
 	})
